requirements.txt
streamlit==1.31.0
httpx==0.26.0
python-dotenv==1.0.0